Update on addition and begining of division
Addition function bug has been fixed. The bug was if the sum of two numbers had a resulting exponent one larger than the biggest and if it was the first number in that expnent's category, it didn't add the exponents.

E.g.
9 + 1 was 1.0
97+ 3 was 10.0

That has been fixed.

P.S. The division function has been started.